+++ lib/app.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:responsive_sizer/responsive_sizer.dart';
+import 'package:rro_web/ui/screen/dashboard/dash_screen.dart';
 
 import 'package:rro_web/utils/colors.dart';
 
-import 'ui/screen/dashboard/dash_desk_screen.dart';
 
 class App extends StatefulWidget {
   const App({Key? key}) : super(key: key);
@@ -16,28 +16,28 @@ class _AppState extends State<App> {
   @override
   Widget build(BuildContext context) {
     return ResponsiveSizer(
+        maxTabletWidth: 700,
         builder: (_, __, ___) => MaterialApp(
-          title: "RRO",
+              title: "RRO",
               theme: ThemeData(
-                fontFamily: 'Proxima',
-                colorScheme: ColorScheme(
-                    primary: AppColors.primary,
-                    secondary: AppColors.primary,
-                    surface: AppColors.primary,
-                    background: AppColors.primary[0]!,
-                    error: systemRedColor,
-                    onBackground: AppColors.primary[0]!,
-                    onSurface: AppColors.primary,
-                    onError: systemRedColor,
-                    onPrimary: AppColors.primary,
-                    onSecondary: AppColors.primary,
-                    brightness: Brightness.light),
-                brightness: Brightness.light,
-                primarySwatch: AppColors.primary,
-                primaryColor: AppColors.primary,
-                scaffoldBackgroundColor: AppColors.primary[100]!
-              ),
-              home: const DashDeskScreen(),
+                  fontFamily: 'Proxima',
+                  colorScheme: ColorScheme(
+                      primary: AppColors.primary,
+                      secondary: AppColors.primary,
+                      surface: AppColors.primary,
+                      background: AppColors.primary[0]!,
+                      error: systemRedColor,
+                      onBackground: AppColors.primary[0]!,
+                      onSurface: AppColors.primary,
+                      onError: systemRedColor,
+                      onPrimary: AppColors.primary,
+                      onSecondary: AppColors.primary,
+                      brightness: Brightness.light),
+                  brightness: Brightness.light,
+                  primarySwatch: AppColors.primary,
+                  primaryColor: AppColors.primary,
+                  scaffoldBackgroundColor: AppColors.primary[100]!),
+              home: const DashScreen(),
             ));
   }
 }