+++ lib/ui/screen/profile/skill/skill_segment.dart
@@ -8,8 +8,9 @@ import 'package:rro_web/utils/texts.dart';
 import '../../../widgets/minor_ui.dart';
 import '../../../widgets/space.dart';
 
-class SkillDeskSegment extends StatelessWidget {
-  const SkillDeskSegment({Key? key}) : super(key: key);
+class SkillSegment extends StatelessWidget {
+  final bool isDesktop;
+  const SkillSegment({Key? key, this.isDesktop = true}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -18,9 +19,8 @@ class SkillDeskSegment extends StatelessWidget {
       crossAxisAlignment: CrossAxisAlignment.start,
       mainAxisAlignment: MainAxisAlignment.start,
       children: [
-        titleSegment("Skill"),
-        Space.vMedium,
-        Space.vMedium,
+        titleSegment("Skill", isDesktop: isDesktop),
+        isDesktop ? Space.vLarge : Space.vMedium,
         _buildItemSegment("Programming Language", skill.proLang),
         Space.vMedium,
         _buildItemSegment("Tools", skill.tools),
@@ -41,19 +41,20 @@ class SkillDeskSegment extends StatelessWidget {
             Icon(
               Icons.arrow_right,
               color: systemPrimaryColor,
-              size: 15.sp,
+              size: isDesktop ? 15.sp : 18.sp,
             ),
             Expanded(
               child: Text(
                 subTitle,
-                style: contentPrimaryBold.copyWith(fontSize: 14.sp),
+                style: contentPrimaryBold.copyWith(
+                    fontSize: isDesktop ? 14.sp : 17.sp),
               ),
             )
           ],
         ),
         Space.vSmall,
         Padding(
-          padding: EdgeInsets.only(left: 15.sp),
+          padding: EdgeInsets.only(left: isDesktop ? 15.sp : 18.sp),
           child: Wrap(
             spacing: 12.0,
             runSpacing: 12.0,
@@ -68,19 +69,20 @@ class SkillDeskSegment extends StatelessWidget {
     return Container(
       padding: const EdgeInsets.symmetric(horizontal: 10, vertical: 8),
       decoration: BoxDecoration(
-        boxShadow: [
-          BoxShadow(
-                  color: AppColors.primary[400]!,
-                  blurRadius: 10,
-                  spreadRadius: 0.01,
-                  offset: const Offset(3, 3),
-                ),
-        ],
+          boxShadow: [
+            BoxShadow(
+              color: AppColors.primary[400]!,
+              blurRadius: 10,
+              spreadRadius: 0.01,
+              offset: const Offset(3, 3),
+            ),
+          ],
           borderRadius: const BorderRadius.all(Radius.circular(5.0)),
           color: systemPrimaryColor),
       child: Text(
         title,
-        style: contentBold.copyWith(color: systemWhiteColor),
+        style: contentBold.copyWith(
+            color: systemWhiteColor, fontSize: isDesktop ? 12.sp : 16.sp),
       ),
     );
   }